requirements.txt
celery==4.2.1
coverage==4.5.1
dj-database-url==0.5.0
Django==2.0.5
django-appconf==1.0.2
django-axes==4.4.0
django-celery-beat==1.5.0
django-cors-headers==2.2.0
django-environ==0.4.5
django-extensions==2.1.4
django-filter==1.1.0
django-ipware==2.1.0
django-redis==4.10.0
django-redis-cache==2.0.0
django-rest-swagger==2.2.0
django-storages==1.6.6
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-autodocs==0.4.4
drf-yasg==1.9.1
factory-boy==2.8.1
Faker==0.9.1
flex==6.13.2
future==0.16.0
gunicorn==19.8.1
httplib2==0.11.3
jws==0.1.3
oauth2client==3.0.0
openapi-codec==1.3.2
pep8==1.7.1
Pillow==5.2.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.3
pylint==1.9.1
pytest==3.1.2
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-pythonpath==0.7.1
pytest-xdist==1.17.1
python-dateutil==2.7.3
python-dotenv==0.9.1
python-http-client==3.1.0
python-jwt==2.0.1
PyYAML==3.13
redis==2.10.5
requests==2.18.0
requests-toolbelt==0.7.0
six==1.11.0
swagger-spec-validator==2.3.1
validate-email==1.3
vine==1.1.4